# .github/backend/backend/Models/Registration.cs
namespace backend.Models
{
    public class Registration
    {
        public int ID { get; set; }

        public string UserName { get; set; }

        public string Password { get; set; }

        public string Email { get; set; }

        public bool IsActive { get; set; }

        public bool IsVerified { get; set; }
    }
}